api/requirements.txt
fastapi
uvicorn
sqlalchemy
loguru
orjson
python-dotenv
flake8
mypy
passlib
bcrypt
PyJWT